rename(route-folder): 알림 페이지 constants 컨벤션 적용
