+++ src/app/(route)/alert/_components/ALERT_CONST.ts
@@ -1,3 +1,24 @@
+export const ALERT_CATEGORIES = [
+  { key: "all", label: "전체" },
+  { key: "category", label: "카테고리 키워드" },
+  { key: "chat", label: "채팅" },
+  { key: "comment", label: "댓글" },
+  { key: "notice", label: "공지사항" },
+  { key: "inquiry_reply", label: "문의" },
+  { key: "report_result", label: "신고" },
+] as const;
+
+export const ALERT_ROW_BG = {
+  read: {
+    default: "bg-white hover:bg-fill-flatGray-25",
+    delete: "bg-white hover:bg-white",
+  },
+  unread: {
+    default: "bg-fill-brand-subtle-default_3 hover:bg-fill-brand-subtle-default_2",
+    delete: "bg-fill-brand-subtle-default_3 hover:bg-fill-brand-subtle-default_3",
+  },
+} as const;
+
 import { NotificationType, ReferenceType } from "@/api/fetch/notification";
 
 type AlertIconBackgroundColor = {
