--- src/app/(route)/alert/_constants/ALERT_CATEGORIES.ts
@@ -1,9 +1,0 @@
-export const ALERT_CATEGORIES = [
-  { key: "all", label: "전체" },
-  { key: "category", label: "카테고리 키워드" },
-  { key: "chat", label: "채팅" },
-  { key: "comment", label: "댓글" },
-  { key: "notice", label: "공지사항" },
-  { key: "inquiry_reply", label: "문의" },
-  { key: "report_result", label: "신고" },
-] as const;
