--- src/app/(route)/alert/_constants/ALERT_ROW_BG.ts
@@ -1,10 +1,0 @@
-export const ALERT_ROW_BG = {
-  read: {
-    default: "bg-white hover:bg-fill-flatGray-25",
-    delete: "bg-white hover:bg-white",
-  },
-  unread: {
-    default: "bg-fill-brand-subtle-default_3 hover:bg-fill-brand-subtle-default_2",
-    delete: "bg-fill-brand-subtle-default_3 hover:bg-fill-brand-subtle-default_3",
-  },
-} as const;
