+++ src/app/(route)/alert/_types/alertKeyType.ts
@@ -1,3 +1,3 @@
-import { ALERT_CATEGORIES } from "../_constants/ALERT_CATEGORIES";
+import { ALERT_CATEGORIES } from "../_components/ALERT_CONST";
 
 export type AlertCategoryKey = (typeof ALERT_CATEGORIES)[number]["key"];
